CRUD operations authorized

+++ BlogApp/controllers/blog.js
@@ -1,7 +1,10 @@
 const { Blog } = require('../models');
+const jwt = require('jsonwebtoken');
 
 const view_all = (req, res) => {
   Blog.find()
+    .where('createdBy')
+    .equals(res.locals.user._id)
     .sort({ createdAt: -1 })
     .then((result) => {
       res.render('blogs/index', { path: req.originalUrl, blogs: result });
@@ -19,12 +22,18 @@ const view_edit = (req, res) => {
   const { blogId } = req.params;
 
   Blog.findById(blogId)
+    .where('createdBy')
+    .equals(res.locals.user._id)
     .then((result) => {
-      res.render('blogs/edit', {
-        path: req.originalUrl,
-        blog: result,
-        pageTitle: 'Edit Blog',
-      });
+      if (result) {
+        res.render('blogs/edit', {
+          path: req.originalUrl,
+          blog: result,
+          pageTitle: 'Edit Blog',
+        });
+      } else {
+        res.redirect('/404');
+      }
     })
     .catch((err) => {
       console.log(err);
@@ -35,12 +44,18 @@ const view_details = (req, res) => {
   const { blogId } = req.params;
 
   Blog.findById(blogId)
+    .where('createdBy')
+    .equals(res.locals.user._id)
     .then((result) => {
-      res.render('blogs/details', {
-        path: req.originalUrl,
-        blog: result,
-        pageTitle: 'Blog Details',
-      });
+      if (result) {
+        res.render('blogs/details', {
+          path: req.originalUrl,
+          blog: result,
+          pageTitle: 'Blog Details',
+        });
+      } else {
+        res.redirect('/404');
+      }
     })
     .catch((err) => {
       console.log(err);
@@ -63,7 +78,11 @@ const api_create = (req, res) => {
 const api_delete = (req, res) => {
   const { blogId } = req.params;
 
+  const { id: userId } = jwt.decode(req.cookies.user_jwt);
+
   Blog.findByIdAndDelete(blogId)
+    .where('createdBy')
+    .equals(userId)
     .then((result) => {
       res.json({ redirect: '/blogs' });
     })
@@ -75,7 +94,11 @@ const api_update = (req, res) => {
   const { body, params } = req;
   const { blogId } = params;
 
+  const { id: userId } = jwt.decode(req.cookies.user_jwt);
+
   Blog.findByIdAndUpdate(blogId, body)
+    .where('createdBy')
+    .equals(userId)
     .then((result) => {
       res.json({
         redirect: `/blogs/${blogId}`,
